--- a/crates/next-core/src/next_app/app_page_entry.rs
+++ b/crates/next-core/src/next_app/app_page_entry.rs
@@ -98,8 +98,8 @@ pub async fn get_app_page_entry(
         fxindexmap! {
             "tree" => loader_tree_code,
             "pages" => StringifyJs(&pages).to_string().into(),
-            "__next_app_require__" => TURBOPACK_REQUIRE.full.into(),
-            "__next_app_load_chunk__" => TURBOPACK_LOAD.full.into(),
+            "__next_app_require__" => TURBOPACK_REQUIRE.bound().into(),
+            "__next_app_load_chunk__" => TURBOPACK_LOAD.bound().into(),
         },
         fxindexmap! {},
     )