--- a/packages/next/src/client/next-turbopack.ts
+++ b/packages/next/src/client/next-turbopack.ts
@@ -29,7 +29,9 @@ initialize({})
       page: string,
       chunksData: any
     ) => {
-      const chunkPromises = chunksData.map(__turbopack_load__)
+      const chunkPromises = chunksData.map((c: unknown) =>
+        __turbopack_load__(c)
+      )
 
       Promise.all(chunkPromises).catch((err) =>
         console.error('failed to load chunks for page ' + page, err)