--- a/turbopack/crates/turbopack-ecmascript-runtime/js/src/browser/runtime/base/dev-base.ts
+++ b/turbopack/crates/turbopack-ecmascript-runtime/js/src/browser/runtime/base/dev-base.ts
@@ -153,26 +153,7 @@ function instantiateModule(
     // This can happen if modules incorrectly handle HMR disposes/updates,
     // e.g. when they keep a `setTimeout` around which still executes old code
     // and contains e.g. a `require("something")` call.
-    let instantiationReason
-    switch (sourceType) {
-      case SourceType.Runtime:
-        instantiationReason = `as a runtime entry of chunk ${sourceData}`
-        break
-      case SourceType.Parent:
-        instantiationReason = `because it was required from module ${sourceData}`
-        break
-      case SourceType.Update:
-        instantiationReason = 'because of an HMR update'
-        break
-      default:
-        invariant(
-          sourceType,
-          (sourceType) => `Unknown source type: ${sourceType}`
-        )
-    }
-    throw new Error(
-      `Module ${id} was instantiated ${instantiationReason}, but the module factory is not available. It might have been deleted in an HMR update.`
-    )
+    factoryNotAvailable(id, sourceType, sourceData)
   }
 
   const hotData = moduleHotData.get(id)!