--- a/turbopack/crates/turbopack-ecmascript-runtime/js/src/browser/runtime/base/dummy.ts
+++ b/turbopack/crates/turbopack-ecmascript-runtime/js/src/browser/runtime/base/dummy.ts
@@ -11,17 +11,3 @@
 /// <reference path="../../../shared/require-type.d.ts" />
 
 declare var BACKEND: RuntimeBackend
-declare var loadWebAssembly: (
-  sourceType: SourceType,
-  sourceData: SourceData,
-  wasmChunkPath: ChunkPath,
-  edgeModule: () => WebAssembly.Module,
-  imports: WebAssembly.Imports
-) => Exports
-declare var loadWebAssemblyModule: (
-  sourceType: SourceType,
-  sourceData: SourceData,
-  wasmChunkPath: ChunkPath,
-  edgeModule: () => WebAssembly.Module
-) => WebAssembly.Module
-declare var relativeURL: (inputUrl: string) => void