--- a/turbopack/crates/turbopack-ecmascript-runtime/js/src/browser/runtime/base/runtime-base.ts
+++ b/turbopack/crates/turbopack-ecmascript-runtime/js/src/browser/runtime/base/runtime-base.ts
@@ -142,6 +142,33 @@ const availableModules: Map<ModuleId, Promise<any> | true> = new Map()
 
 const availableModuleChunks: Map<ChunkPath, Promise<any> | true> = new Map()
 
+function factoryNotAvailable(
+  moduleId: ModuleId,
+  sourceType: SourceType,
+  sourceData: SourceData
+) {
+  let instantiationReason
+  switch (sourceType) {
+    case SourceType.Runtime:
+      instantiationReason = `as a runtime entry of chunk ${sourceData}`
+      break
+    case SourceType.Parent:
+      instantiationReason = `because it was required from module ${sourceData}`
+      break
+    case SourceType.Update:
+      instantiationReason = 'because of an HMR update'
+      break
+    default:
+      invariant(
+        sourceType,
+        (sourceType) => `Unknown source type: ${sourceType}`
+      )
+  }
+  throw new Error(
+    `Module ${moduleId} was instantiated ${instantiationReason}, but the module factory is not available. It might have been deleted in an HMR update.`
+  )
+}
+
 function loadChunk(
   this: TurbopackBrowserBaseContext<Module>,
   chunkData: ChunkData