… test/development/acceptance-app/error-recovery.test.ts
… turbopack/crates/turbopack-ecmascript-runtime/js/src/browser/runtime/dom/runtime-backend-dom.ts
… turbopack/crates/turbopack-ecmascript-runtime/js/src/browser/runtime/edge/runtime-backend-edge.ts
… turbopack/crates/turbopack-ecmascript-runtime/js/src/nodejs/runtime.ts
… turbopack/crates/turbopack-ecmascript-runtime/js/src/shared-node/base-externals-utils.ts
… turbopack/crates/turbopack-ecmascript-runtime/js/src/shared-node/node-externals-utils.ts
… turbopack/crates/turbopack-ecmascript-runtime/js/src/shared/runtime-types.d.ts
… turbopack/crates/turbopack-ecmascript-runtime/js/src/shared/runtime-utils.ts
… turbopack/crates/turbopack-ecmascript/src/chunk/item.rs
… turbopack/crates/turbopack-ecmascript/src/lib.rs
… turbopack/crates/turbopack-ecmascript/src/merged_module.rs
… turbopack/crates/turbopack-ecmascript/src/references/external_module.rs
… turbopack/crates/turbopack-ecmascript/src/references/member.rs
… turbopack/crates/turbopack-ecmascript/src/references/mod.rs
… turbopack/crates/turbopack-ecmascript/src/references/pattern_mapping.rs
… turbopack/crates/turbopack-ecmascript/src/references/unreachable.rs
… turbopack/crates/turbopack-ecmascript/src/runtime_functions.rs
… turbopack/crates/turbopack-ecmascript/src/side_effect_optimization/facade/chunk_item.rs
… turbopack/crates/turbopack-ecmascript/src/side_effect_optimization/locals/chunk_item.rs
… turbopack/crates/turbopack-ecmascript/src/transform/mod.rs
… turbopack/crates/turbopack-ecmascript/src/tree_shake/chunk_item.rs
… turbopack/crates/turbopack-tests/tests/snapshot/basic-tree-shake/dynamic-import/output/4c35f_tests_snapshot_basic-tree-shake_dynamic-import_input_index_92a5f455.js
… turbopack/crates/turbopack-tests/tests/snapshot/basic/async_chunk/output/turbopack_crates_turbopack-tests_tests_snapshot_basic_async_chunk_input_46366300._.js
… turbopack/crates/turbopack-tests/tests/snapshot/basic/async_chunk/output/turbopack_crates_turbopack-tests_tests_snapshot_basic_async_chunk_input_46366300._.js.map
… turbopack/crates/turbopack-tests/tests/snapshot/basic/async_chunk_build/output/4e721_crates_turbopack-tests_tests_snapshot_basic_async_chunk_build_input_1e41378a._.js
… turbopack/crates/turbopack-tests/tests/snapshot/basic/async_chunk_build/output/4e721_crates_turbopack-tests_tests_snapshot_basic_async_chunk_build_input_1e41378a._.js.map
… turbopack/crates/turbopack-tests/tests/snapshot/basic/top-level-await/output/4e721_crates_turbopack-tests_tests_snapshot_basic_top-level-await_input_9acd43f4._.js
… turbopack/crates/turbopack-tests/tests/snapshot/basic/top-level-await/output/4e721_crates_turbopack-tests_tests_snapshot_basic_top-level-await_input_aa0a0c39._.js
… turbopack/crates/turbopack-tests/tests/snapshot/comptime/typeof/output/turbopack_crates_turbopack-tests_tests_snapshot_comptime_typeof_input_80f5724c._.js
… turbopack/crates/turbopack-tests/tests/snapshot/css/css-modules/output/turbopack_crates_turbopack-tests_tests_snapshot_css_css-modules_input_8f2e7f32._.js