@ijjk 2025-07-11
## Failing test suites <!--41571328926-->
Commit: 847a9fb182130df8b31e2592fca72663c57944e5

`pnpm test test/integration/client-404/test/index.test.js` (turbopack)
- Client 404 > development mode > Client Navigation 404 > should show 404 upon client replacestate > should navigate the page

<details>
<summary>Expand output</summary>

  ● Client 404 › development mode › Client Navigation 404 › should show 404 upon client replacestate › should navigate the page

    page.waitForSelector: Timeout 10000ms exceeded.
    Call log:
      - waiting for locator('#hellom8')

      454 |   waitForElementByCss(selector: string, timeout = 10_000) {
      455 |     return this.startChain(async () => {
    > 456 |       const el = await page.waitForSelector(selector, {
          |                             ^
      457 |         timeout,
      458 |         state: 'attached',
      459 |       })

      at waitForSelector (lib/browsers/playwright.ts:456:29)
      at Playwright._chain (lib/browsers/playwright.ts:568:23)
      at Playwright._chain [as startChain] (lib/browsers/playwright.ts:549:17)
      at Playwright.startChain [as waitForElementByCss] (lib/browsers/playwright.ts:455:17)
      at Object.waitForElementByCss (integration/client-404/test/index.test.js:25:23)
      at Proxy._chain (lib/browsers/playwright.ts:568:23)
      at Proxy._chain (lib/browsers/playwright.ts:549:17)
      at Proxy.startChain (lib/browsers/playwright.ts:320:17)
      at Object.back (integration/client-404/test/index.test.js:25:55)

</details>


Read more about building and testing Next.js in [contributing.md](https://github.com/vercel/next.js/blob/canary/contributing.md).